Feat/hito6

+++ src/controllers/project.controller.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from 'express';
 import { EnumUserRole } from '../interfaces/role.interface';
 import { projectService } from '../services/project.service';
-import { getTokenPayload } from '../utils/auth/token.util';
+import { userService } from '../services/user.service';
 
 const getAllProject = async (
     req: Request,
@@ -22,8 +22,8 @@ const getProjectById = async (
     next: NextFunction,
 ) => {
     try {
-        const { id } = req.params;
-        const project = await projectService.getProjectById(id);
+        const { projectId } = req.params;
+        const project = await projectService.getProjectById(projectId);
         res.status(200).json(project);
     } catch (error) {
         next(error);
@@ -36,18 +36,9 @@ const createProject = async (
     next: NextFunction,
 ) => {
     try {
-        const authHeader = req.headers['authorization'];
-
-        if (!authHeader) {
-            res.status(403).json({ error: `Bearer not found` });
-            return;
-        }
-
-        const token = authHeader?.split(' ')[1];
-
-        const { id } = getTokenPayload(token);
         const { name, description } = req.body;
-        console.log(id, name, description, req.body);
+        const { id } = req;
+
         const newProject = await projectService.createProject(
             id,
             name,
@@ -65,7 +56,24 @@ const deleteProject = async (
     next: NextFunction,
 ) => {
     try {
-        const { id } = req.params;
+        const { projectId } = req.params;
+        const { id } = req;
+
+        const projectById = await projectService.getProjectById(projectId);
+
+        if (!projectById) {
+            res.status(404).json({
+                error: `No project found by id ${projectById}`,
+            });
+            return;
+        }
+        if (id !== projectById.created_by && req.role !== EnumUserRole.Admin) {
+            res.status(403).json({
+                error: `Need to be owner or admin to delete the project`,
+            });
+            return;
+        }
+
         const deletedProject = await projectService.deleteProject(id);
         res.status(200).json(deletedProject);
     } catch (error) {
@@ -80,20 +88,18 @@ const updateProject = async (
 ) => {
     try {
         const { name, description, start_date, end_date, status } = req.body;
+        const { id } = req;
 
-        const authHeader = req.headers['authorization'];
-
-        if (!authHeader) {
-            res.status(403).json({ error: `Bearer not found` });
-            return;
-        }
-
-        const token = authHeader?.split(' ')[1];
-
-        const { id } = getTokenPayload(token);
         const { projectId } = req.params;
 
         const projectById = await projectService.getProjectById(projectId);
+
+        if (!projectById) {
+            res.status(404).json({
+                error: `No project found by id ${projectById}`,
+            });
+            return;
+        }
 
         if (id !== projectById.created_by && req.role !== EnumUserRole.Admin) {
             res.status(403).json({
@@ -102,13 +108,7 @@ const updateProject = async (
             return;
         }
 
-        if (!projectById) {
-            res.status(404).json({
-                error: `No project found by id ${projectById}`,
-            });
-            return;
-        }
-        const newProject = await projectService.updateProject(
+        const updatedProject = await projectService.updateProject(
             projectId,
             {
                 name,
@@ -119,13 +119,96 @@ const updateProject = async (
             },
             projectById,
         );
+        res.status(201).json({ updatedProject });
+    } catch (error) {
+        next(error);
+    }
+};
+
+const assingUserToProject = async (
+    req: Request,
+    res: Response,
+    next: NextFunction,
+) => {
+    try {
+        const { userId } = req.body;
+        const { id } = req;
+        const { projectId } = req.params;
+
+        const projectById = await projectService.getProjectById(projectId);
+        const userById = await userService.getUserById(userId);
+
+        if (!projectById) {
+            res.status(404).json({
+                error: `No project found by id ${projectById}`,
+            });
+            return;
+        }
+
+        if (!userById) {
+            res.status(404).json({
+                error: `No user found by id ${userById}`,
+            });
+            return;
+        }
+
+        if (id !== projectById.created_by && req.role !== EnumUserRole.Admin) {
+            res.status(403).json({
+                error: `Need to be owner or admin to update the project`,
+            });
+            return;
+        }
+        const newProject = await projectService.assingUserToProject(
+            userId,
+            projectId,
+        );
         res.status(201).json({ newProject });
     } catch (error) {
         next(error);
     }
 };
 
-// TODO, UPDATE Y DELETE DE PROYECTOS
+const removeUserFromProject = async (
+    req: Request,
+    res: Response,
+    next: NextFunction,
+) => {
+    try {
+        const { userId } = req.body;
+        const { id } = req;
+        const { projectId } = req.params;
+
+        const projectById = await projectService.getProjectById(projectId);
+        const userById = await userService.getUserById(userId);
+
+        if (!projectById) {
+            res.status(404).json({
+                error: `No project found by id ${projectById}`,
+            });
+            return;
+        }
+
+        if (!userById) {
+            res.status(404).json({
+                error: `No user found by id ${userById}`,
+            });
+            return;
+        }
+        if (id !== projectById.created_by && req.role !== EnumUserRole.Admin) {
+            res.status(403).json({
+                error: `Need to be owner or admin to update the project`,
+            });
+            return;
+        }
+        const removeFromProject = await projectService.removeUserFromProject(
+            userId,
+            projectId,
+        );
+        res.status(201).json({ removeFromProject });
+    } catch (error) {
+        next(error);
+    }
+};
 
 export const projectController = {
     getAllProject,
@@ -133,4 +216,6 @@ export const projectController = {
     createProject,
     deleteProject,
     updateProject,
+    assingUserToProject,
+    removeUserFromProject,
 };
